pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "holdfast"
version = "0.1.0"
description = "Governed evolution for prompts and skills — learn from runs, propose bounded improvements, stay within contracts."
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0",
    "jsonschema>=4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.4",
]

[tool.hatch.build.targets.wheel]
packages = ["src/holdfast"]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
